refactor: split lsp/features/semantic_tokens.py into _semantic_tokens/ package
Breaks the 3,216-line module into six focused submodules following the
same pattern as the _analyser/ and _emitter/ splits:

  _constants.py   — token type/modifier lists, keyword sets, core regexes
  _bigip.py       — BIG-IP / APL / iRules token collection
  _primitives.py  — core token emission helpers
  _format_args.py — format/pattern arg-index finders + their collectors
  _collect.py     — main _collect_tokens loop + expression/switch helpers
  _api.py         — public API (semantic_tokens_full, delta encode, etc.)

semantic_tokens.py becomes a ~32-line re-export shim; all 11 callers
(server.py, hover.py, inlay_hints.py, document_state.py, 7 test files)
are unchanged.

_collect_switch_case_bodies is defined in _collect.py (not _format_args.py)
to avoid a circular import with _collect_tokens.
